requirements.txt
pytelegrambotapi==4.17.0
python-tgpt==0.6.6
python-dotenv==1.0.0
click==8.1.3
SQLAlchemy==2.0.29